% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMissingSamples.R
\name{addMissingSamples}
\alias{addMissingSamples}
\title{Add missing samples from GSEs and SRPs within a data frame}
\usage{
addMissingSamples(df, values = c(1, 0))
}
\arguments{
\item{df}{Data frame (must be in searchForAccession 
or searchAnywhere output format)}

\item{values}{A vector (length 2) indicating what will be the values 
assigned to the original df rows and to added rows respectively}
}
\value{
Data frame with added missing rows from the same GSE/SRP. 
It will also contain an additional column to indicate whether the row 
was originally present or not
}
\description{
This function is intended to facilitate workflow with the results of 
\code{searchAnywhere()} function. The main motivation is to provide missing 
samples that did not match the query criteria, but might be useful 
when searching for controls
}
\examples{
startSpiderSeqRDemo()
addMissingSamples()
addMissingSamples(searchForAccession("SRR6823653"))
# NOTE: using searchForAccession to provide the data frame here; 
# in reality the data frame can be generated by most other 
# SpiderSeqR functions (including search functions)



}
\seealso{
Other SpiderSeqR workflow functions: \code{\link{convertAccession}}
}
\concept{SpiderSeqR workflow functions}
