% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendGSEColumns.R
\name{appendGSEColumns}
\alias{appendGSEColumns}
\title{Append information from GSE columns to a data frame}
\usage{
appendGSEColumns(df, gse_columns)
}
\arguments{
\item{df}{Data frame with series_id column containing GSEs; 
gsm column is also required for join purposes}

\item{gse_columns}{Character vector with names of GSE columns to be appended}
}
\value{
Data frame with appended gse columns 
(they will be named with a prepended 'GSE_')
}
\description{
Performs search for GSEs in the gse table and fetches columns of interest, 
which get appended to the existing data frame. In case a sample belongs 
to multiple GSEs, the information appears in the same field separated 
with a special character.
}
\details{
NOTE: pubmed_id column (in gse table from the database generated 
by GEOmetadb) appears to only contain approximately 20% of entries 
from the online version of GEO (the rest has missing pubmed_ids)
}
\examples{
# appendGSEColumns(df, "pubmed_id") # Append only pubmed_id column
# appendGSEColumns(df, "*") # Append all available columns from gse


}
\keyword{internal}
