% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession_Search_Functions.R
\name{batchedAccSearch}
\alias{batchedAccSearch}
\title{Make batched queries to databases}
\usage{
batchedAccSearch(acc_vector, database_name, table_name, col_names,
  c_size = 500)
}
\arguments{
\item{acc_vector}{A character vector with accessions to search for (must belong to the same type)}

\item{database_name}{A character with the name of the database connection}

\item{table_name}{A character with the name of the database table}

\item{col_names}{A character vector with column names to be returned}

\item{c_size}{Number of items to search for in a batch (sqlite has a limit of 999 parameters within a single query)}
}
\value{
Data frame with results of the query
}
\description{
Make batched queries to databases
}
\keyword{internal}
