% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_Functions.R
\name{callRecordGenerator}
\alias{callRecordGenerator}
\title{Generate record of the call}
\usage{
callRecordGenerator(file)
}
\arguments{
\item{file}{A string with the file name (needs to be .Rda to work)}
}
\value{
Nothing. Saves the file with call object
}
\description{
Generate record of the call
}
\keyword{internal}
