% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{checkValidColumns}
\alias{checkValidColumns}
\title{Check if column names of df are within allowed set}
\usage{
checkValidColumns(df)
}
\arguments{
\item{df}{Data frame to be checked}
}
\value{
Nothing - give an error or a warning
}
\description{
Check if column names of df are within allowed set
}
\keyword{internal}
