% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chExtractor.R
\name{chExtractor}
\alias{chExtractor}
\title{Extract characteristics from a vector}
\usage{
chExtractor(df)
}
\arguments{
\item{df}{Data frame with characteristics_ch1 column}
}
\value{
Data frame with extracted information
}
\description{
Extract characteristics from a vector
}
\examples{
# chExtractor(df)

}
\keyword{internal}
