% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession_Functions.R
\name{classifyAccession}
\alias{classifyAccession}
\title{Classify accessions}
\usage{
classifyAccession(x, na.ignore = FALSE)
}
\arguments{
\item{x}{Vector of accessions}

\item{na.ignore}{Logical indicating whether NAs are ignored}
}
\value{
Accession class (error if input does not match any)
}
\description{
Classify accessions
}
\section{Supported accession classes}{

\itemize{
   \item [D|E|S]RP - study_accession
   \item [D|E|S]RS - sample_accession
   \item [D|E|S]RX - experiment_accession
   \item [D|E|S]RR - run_accession
   \item GSE - series_id
   \item GSM - gsm
}

\strong{NOTE:} Input vector needs to contain accessions belonging to only 
one type at any given time, otherwise the function will generate an error.
}

\examples{
# classifyAccession("GSM11111")
# classifyAccession(c("GSM11111", "GSE2222")) # Will throw error



}
\keyword{internal}
