% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{convertLibraryStrategy}
\alias{convertLibraryStrategy}
\title{Convert Between Library Strategy Formats}
\usage{
convertLibraryStrategy(x, input, output, mismatch.ignore = FALSE)
}
\arguments{
\item{x}{Character vector to be converted}

\item{input}{String denoting the input format (see below)}

\item{output}{String denoting the output format (see below)}

\item{mismatch.ignore}{Logical denoting whether mismatches are allowed (if TRUE and no match, original character is returned)}
}
\value{
Library strategy in a desired format

Format types:
\enumerate{
    \item can (canonical) - as exists within the database
    \item short (shorthand) - shortened version of the canonical form
    \item syn (synonyms) - potential synonym
}

Currently, the function supports the following conversions:
\enumerate{
    \item can -> short
    \item syn -> can
}
}
\description{
\code{convertLibraryStrategy} converts between different formats of library strategy strings
}
\keyword{internal}
