% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{convertPairedEnds}
\alias{convertPairedEnds}
\title{Convert the entries to give information about paired ends}
\usage{
convertPairedEnds(df)
}
\arguments{
\item{df}{Data frame to be processed}
}
\value{
Original data frame with a new column with a logical 
indicating whether the run has paired end layout
}
\description{
Convert the entries to give information about paired ends
}
\keyword{internal}
