% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFtsTable.R
\name{createFtsTable}
\alias{createFtsTable}
\title{Create a new table in a dabase to enable fulltext search (fts)}
\usage{
createFtsTable(database_name, table_name, new_table_name)
}
\arguments{
\item{database_name}{Name of database connection}

\item{table_name}{Name of the table to duplicate as fts table}

\item{new_table_name}{Name of the new fts table}
}
\value{
Nothing. Update the relevant database 
(stored in memory or in a '*.sqlite' file)
}
\description{
Creates a new virtual (fts) table with exact same columns 
as the original table, populates it with the contents of original table 
and optimises the table.
}
\examples{
# createFtsTable("sra_con", "sra", "sra_new_ft")


}
\keyword{internal}
