% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{cwt}
\alias{cwt}
\title{Write table with set parameters}
\usage{
cwt(object, filename)
}
\arguments{
\item{object}{Object to be written}

\item{filename}{A character with the name of the file}
}
\value{
Nothing. Write the file
}
\description{
Write table with set parameters
}
\keyword{internal}
