% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{.checkCondition}
\alias{.checkCondition}
\title{Check Condition}
\usage{
.checkCondition(condition, x, message)
}
\arguments{
\item{condition}{String with a regular expression to check within x}

\item{x}{Character vector (or list containing character vectors) 
to be checked}

\item{message}{Error message to be displayed (not required)}
}
\value{
Nothing. Produces an error message if condition 
    is not fulfilled in all vector elements
}
\description{
\code{.checkCondition} verifies a condition (using grepl), 
ignoring any NA elements in a vector
}
\keyword{internal}
