% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.checkDBFile}
\alias{.checkDBFile}
\title{Check and ensure that the DB file is in place}
\usage{
.checkDBFile(path, db_file, db_file_name, db_expiry)
}
\arguments{
\item{path}{Path for storing database files (as passed to startSpiderSeqR())}

\item{db_file}{Path to the database file}

\item{db_file_name}{File name (without the path)}

\item{db_expiry}{Maximum number of days since file was modified}
}
\value{
Nothing. If the file doesn't exist or is out of date, offer to 
    download/create it. Otherwise, print information about the file.
}
\description{
Check and ensure that the DB file is in place
}
