% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp-checkFileDate.R
\name{.checkFileDate}
\alias{.checkFileDate}
\title{Check if DB file is in date (based on metaInfo timestamp)}
\usage{
.checkFileDate(conn, db, sra_expiry, geo_expiry, srr_gsm_expiry)
}
\arguments{
\item{conn}{Database connection}

\item{db}{Database name (one of: SRA, GEO, SpiderSeqR)}

\item{sra_expiry, geo_expiry, srr_gsm_expiry}{Numeric with expiry dates}
}
\value{
A logical indicating whether the file is in date (TRUE) or not 
(FALSE)
}
\description{
Check if DB file is in date (based on metaInfo timestamp)
}
\keyword{internal}
