% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp3-clearup.R
\name{.clearUpDBFiles}
\alias{.clearUpDBFiles}
\title{Remove new files}
\usage{
.clearUpDBFiles(path, previous_files)
}
\arguments{
\item{path}{Path to be searched within}

\item{previous_files}{A character vector with the file list within path}
}
\value{
Nothing. Prints the names of new files and removes them.
}
\description{
Remove new files
}
\keyword{internal}
