% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.createCustomDBFile}
\alias{.createCustomDBFile}
\title{===*===
-------------------------------------------------------
Create custom database file}
\usage{
.createCustomDBFile(sra_file, geo_file)
}
\arguments{
\item{sra_file}{A character with the path to SRA database file}

\item{geo_file}{A character with the path to GEO database file}
}
\value{
Nothing. Create a file with a conversion database 
    between SRA and GEO
}
\description{
===*===
-------------------------------------------------------
Create custom database file
}
\keyword{internal}
