% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqRDemo.R
\name{.createDemoGEO}
\alias{.createDemoGEO}
\title{Create a DEMO GEO database}
\usage{
.createDemoGEO(conn)
}
\arguments{
\item{conn}{A database connection to load the data into}
}
\value{
Nothing. Update the connection object
}
\description{
Create a DEMO GEO database
}
\keyword{internal}
