% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqRDemo.R
\name{.createDemoSRA}
\alias{.createDemoSRA}
\title{Create a DEMO SRA database}
\usage{
.createDemoSRA(conn)
}
\arguments{
\item{conn}{A database connection to load the data into}
}
\value{
Nothing. Update the connection object. 
NOTE: does not currently include the creation of sra_ft table, 
this is done by .createFtsTable("sra_con", "sra", "sra_ft")
}
\description{
Create a DEMO SRA database
}
\keyword{internal}
