% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqRDemo.R
\name{.createDemoSRR_GSM}
\alias{.createDemoSRR_GSM}
\title{Create a DEMO SRR_GSM database}
\usage{
.createDemoSRR_GSM(conn)
}
\arguments{
\item{conn}{A database connection to load the data into}
}
\value{
Nothing. Update the connection object
}
\description{
Create a DEMO SRR_GSM database
}
\keyword{internal}
