% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.createMockGEO}
\alias{.createMockGEO}
\title{Create a mock GEO database file (for use in testing)}
\usage{
.createMockGEO(path)
}
\arguments{
\item{path}{A character with the path to the directory for the database file}
}
\value{
Nothing. Creates the database file (unless already present) 
    and writes relevant tables.
}
\description{
Create a mock GEO database file (for use in testing)
}
\keyword{internal}
