% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.createMockSRA}
\alias{.createMockSRA}
\title{-------------------------------------------------------
===*====
Suggested names:
createMockSRA
createMOckGEO
createMockCustomDB}
\usage{
.createMockSRA(path)
}
\arguments{
\item{path}{A character with the path to the directory for the database file}
}
\value{
Nothing. Creates the database file (unless already present) 
    and writes relevant tables.
}
\description{
Course of action
- fetch the tables from the environment
- set up the database connection into the relevant file (?path)
- write the tables into the database
- close the connection
- make sure that the directory is changed back to original
}
\details{
Options:
- everything in one function per DB
- split into db specific part and writing dfs into the database


.createDBFile
- df
- name
- database file name
- path to database file name

- path to file
- (file known)
- (tables known)

Create a mock SRA database file (for use in testing)
}
\keyword{internal}
