% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.createSpiderMetaInfo}
\alias{.createSpiderMetaInfo}
\title{Create metaInfo table for SpiderSeqR database}
\usage{
.createSpiderMetaInfo(sra_file, geo_file)
}
\arguments{
\item{sra_file}{A character to the path with SRA file}

\item{geo_file}{A character to the path with GEO file}
}
\value{
A dataframe with metaInfo
}
\description{
Create metaInfo table for SpiderSeqR database
}
\keyword{internal}
