% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{.detectControls}
\alias{.detectControls}
\title{Detect controls in a sample sheet}
\usage{
.detectControls(df)
}
\arguments{
\item{df}{Data frame to be checked (with a subset of sra columns)}
}
\value{
Data frame with control column entries labelled as 'control' 
where appropriate
}
\description{
Detect controls in a sample sheet
}
\section{Logic of control detection}{

\itemize{
    \item NECESSARY - these string matches are required 
    (entries not fulfilling all necessary conditions 
        will not be considered for labelling)
    \item ANTIBODY - rows with non-empty sa_antibody column 
        will only be labelled based on that field
    \item MATCH - one of these string matches are sufficient 
        to label an entry (provided it fulfils NECESSARY conditions 
        and that it has an EMPTY (NA) sa_antibody field)
    \item OTHERWISE - one of these string matches is sufficient 
        to label an entry, but ONLY if none of the SRP members 
        is labelled as input
}
}

\keyword{internal}
