% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{.detectInputs}
\alias{.detectInputs}
\title{Detect inputs in a sample sheet}
\usage{
.detectInputs(df)
}
\arguments{
\item{df}{Data frame to be checked (with a subset of sra columns)}
}
\value{
Data frame with input column entries labelled as 'input' 
where appropriate
}
\description{
Detect inputs in a sample sheet
}
\section{Logic of input detection}{


The entries are labelled as inputs based on the following rules. 
They strictly need to fulfill NECESSARY conditions as well as:
\itemize{
    \item 1A: have matches to ANTIBODY conditions
    \item 1B: have matches to MATCH conditions, 
          provided no ANTIBODY information was provided
    \item 2: the whole SRP had no matches to ANTIBODY or MATCH conditions, 
          but there is a match to OTHERWISE conditions
}


Further details on the conditions:
 

\itemize{
    \item NECESSARY - these string matches are required 
    (entries not fulfilling all necessary conditions 
        will not be considered for labelling)
    \item ANTIBODY - rows with non-empty sa_antibody column 
        will only be labelled based on that field
    \item MATCH - one of these string matches are sufficient 
        to label an entry (provided it fulfils NECESSARY conditions 
        and that it has an EMPTY (NA) sa_antibody field)
    \item OTHERWISE - one of these string matches is sufficient 
        to label an entry, but ONLY if none of the SRP members 
        is labelled as input
}
}

\keyword{internal}
