% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_Anywhere.R
\name{.findAccessionLevelColumnNames}
\alias{.findAccessionLevelColumnNames}
\title{Find column names corresponding to accession levels in SRA and GEO (SP)}
\usage{
.findAccessionLevelColumnNames(
  acc_levels = c("run", "experiment", "sample", "gsm")
)
}
\arguments{
\item{acc_levels}{Accession levels}
}
\value{
Vector with column names

NOTE: works on the SpiderSeqR column names (i.e. with prefixes)
}
\description{
Find column names corresponding to accession levels in SRA and GEO (SP)
}
\examples{
# .findAccessionLevelColumnNames("run")

}
\keyword{internal}
