% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.findDBFiles}
\alias{.findDBFiles}
\title{Functions:
 - .findDBFiles
- .missingFileCheck
- .setExpiryParameters
- .DBNames
- .checkDBNames
- .checkDBFile
- .noDBFile
- .oldDBFile
- .DBFileExists
- .getDBFile
- .createCustomDBFile
- .getFurtherDBInfo
Find database files}
\usage{
.findDBFiles(path)
}
\arguments{
\item{path}{Path to search within}
}
\value{
A list with paths to database files
}
\description{
Functions:
 - .findDBFiles
- .missingFileCheck
- .setExpiryParameters
- .DBNames
- .checkDBNames
- .checkDBFile
- .noDBFile
- .oldDBFile
- .DBFileExists
- .getDBFile
- .createCustomDBFile
- .getFurtherDBInfo
Find database files
}
\keyword{internal}
