% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.findDBFiles_Ori}
\alias{.findDBFiles_Ori}
\title{Find database files}
\usage{
.findDBFiles_Ori(path)
}
\arguments{
\item{path}{Path to search within}
}
\value{
A list with paths to database files
}
\description{
Find database files
}
\keyword{internal}
