% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR.R
\name{.findFiles_Ori}
\alias{.findFiles_Ori}
\title{Find files (wrapper around dir allowing to search a few levels up)}
\usage{
.findFiles_Ori(pattern, recurse_levels)
}
\arguments{
\item{pattern}{Regular expression pattern to search for (passed to dir())}

\item{recurse_levels}{Integer with the number of directory levels 
(above current working directory) to search in}
}
\value{
A full path with the matching file(s)
}
\description{
Find files (wrapper around dir allowing to search a few levels up)
}
\examples{
#.findFiles("*.sqlite", 5)

}
\keyword{internal}
