% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{.generateDBExtract}
\alias{.generateDBExtract}
\title{Generate DB extract}
\usage{
.generateDBExtract(df)
}
\arguments{
\item{df}{Data frame for export}
}
\value{
A data frame with only relevant columns for export
}
\description{
The columns for the DB extract can be viewed or set 
by \code{\link{getSpiderSeqROption}} and \code{\link{setSpiderSeqROption}}
respectively.
}
\keyword{internal}
