% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_Functions.R
\name{.generateFileName_CALL_SA}
\alias{.generateFileName_CALL_SA}
\title{Generate file name for call record (searchAnywhere)}
\usage{
.generateFileName_CALL_SA(x)
}
\arguments{
\item{x}{A list containing relevant parameters from searchAnywhere}
}
\value{
A string with the file name
}
\description{
Generate file name for call record (searchAnywhere)
}
\keyword{internal}
