% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_Functions.R
\name{.generateFileName_CALL_SFA}
\alias{.generateFileName_CALL_SFA}
\title{Generate call file for searchForAccession}
\usage{
.generateFileName_CALL_SFA(acc_vector)
}
\arguments{
\item{acc_vector}{A character vector with accessions searched for}
}
\value{
A character vector with the file name

Current format:

SFA_ACC1_ACCN_Nn_d-a-te.Rda
}
\description{
Generate call file for searchForAccession
}
\keyword{internal}
