% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_Functions.R
\name{.generateParameterRecord}
\alias{.generateParameterRecord}
\title{Generate a record of parameters}
\usage{
.generateParameterRecord(st, file, fun_name)
}
\arguments{
\item{st}{A list of arguments}

\item{file}{A string with the file name}

\item{fun_name}{A string with the function name}
}
\value{
Nothing. Create a file with the record
}
\description{
Generate a record of parameters
}
\keyword{internal}
