% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{.generateUniversalSampleSheet}
\alias{.generateUniversalSampleSheet}
\title{Generate a sample sheet (universal function)}
\usage{
.generateUniversalSampleSheet(df, SRA_library_strategy)
}
\arguments{
\item{df}{Data frame for export}

\item{SRA_library_strategy}{A character vector with library strategy 
(used for determining the type of sample sheet to be generated)}
}
\value{
A data frame with only relevant columns for export
}
\description{
A wrapper function for selecting and creating 
    the relevant sample sheet type for the data provided
}
\keyword{internal}
