% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.getFurtherDBInfo}
\alias{.getFurtherDBInfo}
\title{Get futher database information (metaInfo)}
\usage{
.getFurtherDBInfo(db_file_name, database_name)
}
\arguments{
\item{db_file_name}{A character with the name of the database file}

\item{database_name}{A character with the database (connection) name}
}
\value{
Nothing. Display messages with information in the console
}
\description{
Get futher database information (metaInfo)
}
\keyword{internal}
