% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{.manageLibraryStrategy}
\alias{.manageLibraryStrategy}
\title{Manage Library Strategy Formats}
\usage{
.manageLibraryStrategy(x, input, output, task = "conv",
  mismatch.ignore = FALSE)
}
\arguments{
\item{x}{Character vector to be converted}

\item{input}{String denoting the input format (see below)}

\item{output}{String denoting the output format (see below)}

\item{task}{String denoting the task to be performed (see below)}

\item{mismatch.ignore}{Logical denoting whether mismatches are allowed 
(if TRUE and no match, original character is returned)}
}
\value{
Library strategy in a desired format

Format types:
\enumerate{
    \item can (canonical) - as exists within the database
    \item short (shorthand) - shortened version of the canonical form
    \item syn (synonyms) - potential synonym
}


Available tasks:
\enumerate{
    \item conv - convert between formats
    \item ex - produce a list of library strategies
    \item check_can - check if x is in canonical form
}


Currently, the function supports the following conversions:
\enumerate{
    \item can -> short, i.e. input = "can", ouput = "short"
    \item syn -> can, i.e. input = "syn", output = "can"
}
}
\description{
\code{.manageLibraryStrategy} by default converts between different formats 
of library strategy strings and offers a few other tasks 
related to library strategies
}
\examples{
# Convert into short form
# .manageLibraryStrategy("RNA-Seq", input = "can", output = "short") 

# Convert into canonical form
# .manageLibraryStrategy("RNA", input = "syn", output = "can") 

# List supported formats for library strategy
# .manageLibraryStrategy(task = "ex") 

# Check whether library strategy is in canonical form
# .manageLibraryStrategy("RNA-Seq", task = "check_can") 




}
\keyword{internal}
