% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.missingFileCheck}
\alias{.missingFileCheck}
\title{Check for missing files}
\usage{
.missingFileCheck(
  sra_file = sra_file,
  geo_file = geo_file,
  srr_gsm_file = srr_gsm_file
)
}
\arguments{
\item{sra_file}{Path to SRA file}

\item{geo_file}{Path to GEO file}

\item{srr_gsm_file}{Path to SRR_GSM file}
}
\value{
A logical vector length 3 indicating whether the respective 
    files are present (in order as above)
}
\description{
Check for missing files
}
\keyword{internal}
