% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.noDBFile}
\alias{.noDBFile}
\title{Course of action to follow if a db file is missing}
\usage{
.noDBFile(path, db_file_name)
}
\arguments{
\item{path}{Path for storing database files}

\item{db_file_name}{A character with the db_file_name (SRAmetadb.sqlite, 
GEOmetadb.sqlite or SRR_GSM.sqlite)}
}
\value{
Nothing. Offer to download/create the file, otherwise return error.
}
\description{
Course of action to follow if a db file is missing
}
\keyword{internal}
