% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unifying_Functions.R
\name{.orderDFColumns}
\alias{.orderDFColumns}
\title{Order columns in a data frame}
\usage{
.orderDFColumns(df)
}
\arguments{
\item{df}{Data frame with columns to be ordered}
}
\value{
Data frame with ordered columns

The columns are matched against the list in listValidColumns() 
(in the order of subsets). Any non-matching columns are not included
}
\description{
Order columns in a data frame
}
\keyword{internal}
