% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR.R
\name{.progressBar}
\alias{.progressBar}
\title{Generate a progress bar scaled by width}
\usage{
.progressBar(i, n, width = 78)
}
\arguments{
\item{i}{Current stage}

\item{n}{Total number of stages}

\item{width}{Number of characters for progress display 
(the whole bar will have 2 extra ones used for marking both ends)}
}
\value{
Nothing. Output progress to console
}
\description{
Generate a progress bar scaled by width
}
\examples{

#for (i in 1:10){
#    .progressBar(i, 10, 100)
#    Sys.sleep(1)
#}

}
\keyword{internal}
