% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_Functions.R
\name{.replaceForbiddenCharacters}
\alias{.replaceForbiddenCharacters}
\title{Replace forbidden characters from a string}
\usage{
.replaceForbiddenCharacters(x)
}
\arguments{
\item{x}{A character vector with characters to be replaced}
}
\value{
Character vector with replaced forbidden characters
}
\description{
Replace forbidden characters from a string
}
\keyword{internal}
