% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saExtractor.R
\name{.saExtractor}
\alias{.saExtractor}
\title{Extract sample attributes}
\usage{
.saExtractor(df)
}
\arguments{
\item{df}{Data frame with (SRA_)sample_attribute column}
}
\value{
Data frame with extracted information 
    from the sample attribute column and a few added new columns, 
    currently: sa_remainder, sa_tissue, sa_antibody, sa_gene, sa_treatment
}
\description{
This function is a wrapper around .universalExtractor 
with key words specific for (SRA_)sample_attribute field
}
\keyword{internal}
