% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_Anywhere.R
\name{.searchAnywhereGEO}
\alias{.searchAnywhereGEO}
\title{Search anywhere within gse and gsm tables of GEO database}
\usage{
.searchAnywhereGEO(
  GSM_query,
  GSE_query,
  acc_levels = c("gse", "gsm"),
  GEO_type = NULL
)
}
\arguments{
\item{GSM_query}{String to search for within gsm table}

\item{GSE_query}{String to search for within gse table}

\item{acc_levels}{Character vector indicating where to conduct the search 
(only "gse" and "gsm" are considered)}

\item{GEO_type}{Study type for filtering results (optional)}
}
\value{
Data frame with result data from whole GEO (gsm and gse tables)
}
\description{
Search anywhere within gse and gsm tables of GEO database
}
\keyword{internal}
