% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{.searchForSRPChildren}
\alias{.searchForSRPChildren}
\title{Search for missing SRRs within SRPs}
\usage{
.searchForSRPChildren(srp_list, srp_columns)
}
\arguments{
\item{srp_list}{Character vector with SRPs to search for}

\item{srp_columns}{Character vector with names of columns 
(within sra table of SRA) to be retrieved from the database
NOTE: can also be "*", i.e. all columns}
}
\value{
Data frame with all the SRRs belonging to specified SRPs
}
\description{
Search for SRPs within SRA database to retrieve 
their contents (i.e. all SRRs) as a data frame.
}
\keyword{internal}
