% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{.searchSRA}
\alias{.searchSRA}
\title{An internal function for searchForTerm}
\usage{
.searchSRA(SRA_library_strategy, gene, antibody, cell_type, treatment,
  species, platform)
}
\arguments{
\item{SRA_library_strategy}{Character vector specifying experiment approach}

\item{gene}{...and...}

\item{antibody}{...and...}

\item{cell_type}{...and...}

\item{treatment}{...and...}

\item{species}{...and...}

\item{platform}{Character vectors 
with information to search within SRA}
}
\value{
A data frame with results
}
\description{
An internal function for searchForTerm
}
\keyword{Test'}
\keyword{effects:}
\keyword{internal}
\keyword{the}
\keyword{to}
\keyword{visualise}
