% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession_Search_Functions.R
\name{.searchSRR_GSM}
\alias{.searchSRR_GSM}
\title{Search within SRR_GSM database}
\usage{
.searchSRR_GSM(acc_vector, srr_gsm_columns = c("run_accession", "gsm"))
}
\arguments{
\item{acc_vector}{A character vector with accession numbers (GSM or SRR)}

\item{srr_gsm_columns}{A character vector with columns to be returned 
(from within srr_gsm table)}
}
\value{
A data frame with results
}
\description{
Search within SRR_GSM database
}
\examples{
# startSpiderSeqRDemo()
# .searchSRR_GSM("GSM1173367")


}
\keyword{internal}
