% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR.R
\name{.setDBConnections}
\alias{.setDBConnections}
\title{Set database connections}
\usage{
.setDBConnections(sra_file, geo_file, srr_gsm_file)
}
\arguments{
\item{sra_file}{SRA file name (including path)}

\item{geo_file}{GEO file name (including path)}

\item{srr_gsm_file}{SRR_GSM file name (including path)}
}
\value{
Set up database connections in the global environment
}
\description{
Set database connections
}
