% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.setExpiryParameters}
\alias{.setExpiryParameters}
\title{Set expiry parameters}
\usage{
.setExpiryParameters(
  general_expiry,
  sra_expiry,
  geo_expiry,
  srr_gsm_expiry,
  missing_file_number
)
}
\arguments{
\item{general_expiry}{Maximum number of days since database 
file modification}

\item{sra_expiry, geo_expiry, srr_gsm_expiry}{Maximum number of days since 
modification of respective database files}

\item{missing_file_number}{Number of missing files}
}
\value{
Expiry parameters

Sets expiry parameters for three respective databases, according to the 
logic that specific parameters should be used where possible and for missing
parameters general_expiry will be used (i.e. if all arguments are provided,
general_expiry will be ignored).
Also checks whether the expiry parameters are numeric and returns 
appropriate errors or warnings.
}
\description{
Set expiry parameters
}
\keyword{internal}
