% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{.tmenu}
\alias{.tmenu}
\title{Some function that will use .tmenu
Menu function with testing interface}
\usage{
.tmenu(choices, menu_name, title = NULL)
}
\arguments{
\item{choices}{A character vector of choices}

\item{menu_name}{A character with the menu name 
(for access of pre-set options for testing, under 'test_input' option)}

\item{title}{Menu title (defaults to NULL)}
}
\value{
Chosen menu option, either by the users, or by accessing 
pre-set options
}
\description{
Some function that will use .tmenu
Menu function with testing interface
}
\keyword{internal}
