% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{.unifyNAs}
\alias{.unifyNAs}
\title{Unify handling of NA values}
\usage{
.unifyNAs(x)
}
\arguments{
\item{x}{data frame}
}
\value{
Data frame with "" and "NA" values converted to NA
}
\description{
Unify handling of NA values
}
