% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{.verifyMissingRuns}
\alias{.verifyMissingRuns}
\title{Verify missing Runs}
\usage{
.verifyMissingRuns(srr_list_in)
}
\arguments{
\item{srr_list_in}{A character vector with SRRs}
}
\value{
Nothing but a printed notice on whether the SRXs 
to which the SRRs belong also have any other SRRs
}
\description{
Verify missing Runs
}
\keyword{internal}
