% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{.withOut}
\alias{.withOut}
\title{Remove matching elements from vector}
\usage{
.withOut(names, vector)
}
\arguments{
\item{names}{Character vector with values to be removed}

\item{vector}{Character vector from which elements matching 
to names will be removed}
}
\value{
Original vector .withOut specified elements
}
\description{
Remove matching elements from vector
}
\keyword{internal}
