% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{.writeQuery}
\alias{.writeQuery}
\title{Automate Writing SQLite Queries}
\usage{
.writeQuery(term, fields, sql_before = " LIKE '\%", sql_after = "\%'")
}
\arguments{
\item{term}{Character vector with strings of interest}

\item{fields}{character vector with columns of interest}

\item{sql_before}{SQLite query string before term}

\item{sql_after}{SQLite query string after term}
}
\value{
Part of the  SQLite query pertaining to row selection criteria 
(i.e. that comes after specifying the table and which columns to select)
}
\description{
Automate Writing SQLite Queries
}
\keyword{internal}
