% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSpiderSeqR-Helper_Functions.R
\name{.writeTableToFile}
\alias{.writeTableToFile}
\title{Write table to database file}
\usage{
.writeTableToFile(df, table_name, path, database_file, overwrite = FALSE)
}
\arguments{
\item{df}{Data frame to be written as a table in the database}

\item{table_name}{A character with the name for the table}

\item{database_file}{A character with the name of the database file 
(without the path)}

\item{overwrite}{A logical indicating whether to overwrite an existing 
table with the same name. Defaults to FALSE.}
}
\value{
Nothing. Write the dataframe into the specified database
}
\description{
Write table to database file
}
\keyword{internal}
