% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{extractGSM}
\alias{extractGSM}
\title{Extract GSM from the SRA_experiment_title column}
\usage{
extractGSM(output_list, sampleColumn = TRUE)
}
\arguments{
\item{output_list}{Data frame with (SRA_)experiment_title column}

\item{sampleColumn}{Logical indicating whether a new column 
should be added with extracted GSM numbers}
}
\value{
Original data frame with extracted GSMs from (SRA_)experiment_title
column and (if sampleColumn is set to TRUE) a new column 
with extracted GSM numbers
}
\description{
Extract GSM from the SRA_experiment_title column
}
\keyword{internal}
