% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_Anywhere.R
\name{filterByTermByAccessionLevel}
\alias{filterByTermByAccessionLevel}
\title{Filter df according to query matches 
only within accession levels of interest}
\usage{
filterByTermByAccessionLevel(query, df, acc_levels = c("run",
  "experiment", "sample", "gsm"))
}
\arguments{
\item{query}{Query to be passed to MATCH operator (for fts)}

\item{df}{Data frame to be filtered}

\item{acc_levels}{Accession levels to search within 
(choose from: run, experiment, sample, study, gsm, gse); 
defaults to c("run", "experiment", "sample", "gsm")}
}
\value{
Filtered df (containing only rows matching query 
    within specified accession levels)
}
\description{
Performs fts search on the data frame according to the query, only searching
in the columns corresponding to specified accession levels of interest
}
\seealso{
Other Workflow functions: \code{\link{addMissingSamples}},
  \code{\link{convertAccession}},
  \code{\link{filterByTerm}},
  \code{\link{searchForAccession}}

Other Manipulation functions: \code{\link{filterByTerm}},
  \code{\link{orderDFAccessions}}
}
\concept{Manipulation functions}
\concept{Workflow functions}
