% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_Anywhere.R
\name{findSRAAccessionLevelColumnNames}
\alias{findSRAAccessionLevelColumnNames}
\title{Find SRA column names corresponding to accession levels}
\usage{
findSRAAccessionLevelColumnNames(acc_levels = c("run", "experiment",
  "sample", "study"), add_run_accession = TRUE, table_name = "sra_ft")
}
\arguments{
\item{acc_levels}{Accession levels}

\item{add_run_accession}{Logical indicating whether to add run_accession column name}

\item{table_name}{A character with table name}
}
\value{
Vector with column names
}
\description{
Find SRA column names corresponding to accession levels
}
\examples{
findSRAAccessionLevelColumnNames("run")


}
