% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{generateEmptyDF}
\alias{generateEmptyDF}
\title{Generate empty df with columns corresponding to database columns}
\usage{
generateEmptyDF(tables = c("sra", "gsm", "gse", "added"))
}
\arguments{
\item{tables}{Character vector with tables from which the columns}
}
\value{
Data frame with columns corresponding to database columns (with names prepended with appropriate prefix). Format corresponds to that of searchForAccessionAcrossDBsDF
}
\description{
Generate empty df with columns corresponding to database columns
}
\keyword{internal}
