% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{generateFileName}
\alias{generateFileName}
\title{Generate a file name}
\usage{
generateFileName(SRA_library_strategy, gene = NULL, antibody = NULL,
  cell_type = NULL, treatment = NULL, species = NULL,
  platform = NULL, SRA_secondary_library_strategy = NULL, output,
  file_type)
}
\arguments{
\item{SRA_library_strategy, gene, antibody, cell_type}{and:}

\item{treatment, species, platform, SRA_secondary_libarary_strategy}{Strings 
with search parameters used to generate the data}

\item{output}{String with output type}

\item{file_type}{String with file type}
}
\value{
A string with the file name
}
\description{
Generate a file name
}
\keyword{internal}
