% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{generateFileName_Accession}
\alias{generateFileName_Accession}
\title{Generate a file name (Accession)}
\usage{
generateFileName_Accession(SRA_library_strategy, accession, output,
  file_type)
}
\arguments{
\item{SRA_library_strategy}{A string with the library strategy}

\item{accession}{A string with the accesion searched for (will be truncated)}

\item{output}{String with output type}

\item{file_type}{String with file type}
}
\value{
A string with the file name
}
\description{
Generate a file name (Accession)
}
\keyword{internal}
