% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{generateOutput}
\alias{generateOutput}
\title{Generate File Output}
\usage{
generateOutput(df, ss = NULL, st)
}
\arguments{
\item{df}{Data frame to output (unfilter)}

\item{ss}{A character vector with GSEs from verifySuperseries}

\item{st}{search terms (used for conditional filtering and file naming)}
}
\value{
Nothing (generates relevant files)
}
\description{
Generate File Output
}
\details{
Orders the df, replaces tabs (in characteristics_ch1), 
filters rows as appropriate, outputs files.
}
\section{Files created by generateOutput}{

 \itemize{
    \item TARGET - samples of interest with relevant 
        inputs/controls (ChIP/RNA)
    \item ALL - all samples within SRP (that belong to the 
        SRA_library_strategy) - only in the case of ChIP/RNA
    \item SECONDARY - samples that belong to SRA_secondary_library_strategy
        - only for ChIP with specified library_strategy
}
All the files are made available as sample sheets (.csv and .Rda) 
and database extracts (.csv, .Rda)
}

\keyword{internal}
