% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{generateSampleSheet_ChIP}
\alias{generateSampleSheet_ChIP}
\title{Generate ChIP sample sheet}
\usage{
generateSampleSheet_ChIP(df)
}
\arguments{
\item{df}{Data frame for export}
}
\value{
A data frame with only relevant columns for export 
    (compatible with the pipeline)
}
\description{
Generate ChIP sample sheet
}
\keyword{internal}
