% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{geoFinder}
\alias{geoFinder}
\title{Search for GSMs in GEO}
\usage{
geoFinder(db_con, gsm_list, gsm_columns, gse_columns)
}
\arguments{
\item{db_con}{Database connection (object, not character)}

\item{gsm_list}{GSMs to search for}

\item{gsm_columns}{Columns within gsm table}

\item{gse_columns}{Columns within gse table

Performs search within GEO gsm and gse tables}
}
\description{
Search for GSMs in GEO
}
\keyword{internal}
