% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_Option_Settings.R
\name{getSpiderSeqROption}
\alias{getSpiderSeqROption}
\title{Inspect the value of package options}
\usage{
getSpiderSeqROption(name)
}
\arguments{
\item{name}{A character denoting name of the variable}
}
\value{
Value of the variable
}
\description{
Inspect the value of package options
}
\examples{
getSpiderSeqROption("file_output")

}
