% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gse_demo}
\alias{gse_demo}
\title{DEMO extract from gse table (from sqlite file of GEOmetadb package)}
\format{An object of class \code{data.frame} with 13 rows and 18 columns.}
\usage{
gse_demo
}
\description{
Series included in the data frame:
\itemize{
 \item GSE48253
 \item GSE69001
 \item GSE27360
 \item GSE82246
 \item GSE36467
 \item GSE76553
 \item GSE10309
 \item GSE80767
 
}
}
\details{
A data frame created as a DEMO for gse table from sqlite file of GEOmetadb package. Contains a few arbitrarily chosen complete series (including superseries and subseries when appropriate).
}
\keyword{datasets}
