% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsm_demo}
\alias{gsm_demo}
\title{DEMO extract from gsm table (from sqlite file of GEOmetadb package)}
\format{An object of class \code{data.frame} with 111 rows and 32 columns.}
\usage{
data(gsm_demo)
}
\description{
Series included in the data frame:
\itemize{
 \item GSE48253
 \item GSE69001
 \item GSE27360
 \item GSE82246
 \item GSE36467
 \item GSE76553
 \item GSE10309
 \item GSE80767
 
}
}
\details{
A data frame created as a DEMO for gsm table from sqlite file of GEOmetadb package. Contains all samples from a few arbitrarily chosen series (including superseries and subseries when appropriate).
}
\keyword{datasets}
