% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession_Search_Functions.R
\name{listGSEFields}
\alias{listGSEFields}
\title{List all column names of the GSE table}
\usage{
listGSEFields(omit_gse = TRUE)
}
\arguments{
\item{omit_gse}{Logical indicating whether 
to omit 'gse' column name from the output}
}
\value{
Character vector containing column names of GSE table (except gse)
}
\description{
List all column names of the GSE table
}
\examples{
listGSEFields()
 
}
\keyword{internal}
