% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession_Search_Functions.R
\name{listGSMFields}
\alias{listGSMFields}
\title{List all column names of the GSM table}
\usage{
listGSMFields()
}
\value{
Character vector containing column names of GSM table
}
\description{
List all column names of the GSM table
}
\examples{
listGSMFields()
 
}
\keyword{internal}
