% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession_Search_Functions.R
\name{listSRAFields}
\alias{listSRAFields}
\title{List all column names of SRA table}
\usage{
listSRAFields()
}
\value{
Character vector containing column names of SRA table
}
\description{
List all column names of SRA table
}
\examples{
# listSRAFields()
 
}
\keyword{internal}
