% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unifying_Functions.R
\name{listValidColumns}
\alias{listValidColumns}
\title{List valid columns within the data frames}
\usage{
listValidColumns()
}
\value{
A list with column names grouped into categories
}
\description{
List valid columns within the data frames
}
\section{Column Categories}{

The following categories are available:
\itemize{
    \item SRA - columns from the sra table (SRA database) 
        with added 'SRA_' prefix
    \item GSM - columns from the gsm table (GEO database) 
        with added 'GSM_' prefix
    \item GSE - columns from the gse table (GSE database) 
        with added 'GSE_' prefix
    \item Other - other columns (created within the pacakge) 
        with added 'OTH_' prefix
}
}

\examples{
startSpiderSeqRDemo()
listValidColumns() # List all columns
listValidColumns()$SRA # List columns from sra table
listValidColumns()$GSM # List columns from gsm table
listValidColumns()$GSE # List columns from gse table
listValidColumns()$Other # List other columns

}
