% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{manageColumns}
\alias{manageColumns}
\title{Manage columns within a data frame}
\usage{
manageColumns(df, df_columns, out_columns)
}
\arguments{
\item{df}{Data frame}

\item{df_columns}{A character vector with column names 
to be selected from the data frame}

\item{out_columns}{A character vector with new column names}
}
\value{
A data frame with only specified columns 
    (renamed as specified in out_columns).
}
\description{
Manage columns within a data frame
}
\keyword{internal}
