% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Console_Messages.R
\name{mm}
\alias{mm}
\title{Console Messages}
\usage{
mm(string, type)
}
\arguments{
\item{string}{A character to be displayed}

\item{type}{A type of message}
}
\value{
Console message
}
\description{
Console Messages
}
\keyword{internal}
