% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{orderAccessions}
\alias{orderAccessions}
\title{Order Accessions}
\usage{
orderAccessions(x, na.last = TRUE)
}
\arguments{
\item{x}{Character vector (or list containing character vectors) to be ordered}

\item{na.last}{For controlling the treatment of NAs (see \code{\link[base]{order}})}
}
\value{
Integer vector with order of indices
}
\description{
\code{orderAccessions} orders character vectors or lists according to numbers, disregarding any lettrs.
}
\examples{
\dontrun{
# Order rows in df according to multiple accession types
order_columns <- list(df$study_accession, df$sample_accession, df$experiment_accession, df$run_accession, df$gsm)
df <- df[orderAccessions(order_columns), ]
}

}
\seealso{
\code{\link[base]{order}}
}
