% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unifying_Functions.R
\name{orderDFAccessions}
\alias{orderDFAccessions}
\title{Order df rows in accession order}
\usage{
orderDFAccessions(df, acc_order = c("study", "sample", "experiment",
  "run", "gsm"), na.last = TRUE)
}
\arguments{
\item{df}{Data frame to be ordered (must contain columns corresponding 
to types specified in \code{acc_order}, i.e. study_accession, 
sample_accession, experiment_accession, run_accession, gsm).}

\item{acc_order}{A character vector with accession levels to be used 
for ordering (their order specifies which accession level is taken 
into account first). Default value lists all possible elements.}

\item{na.last}{Logical denoting whether NA values should be placed last. 
Defaults to TRUE.}
}
\value{
Data frame with rows ordered accordingly. The row names are reset.
}
\description{
Order df rows in accession order
}
\examples{
# Use for ordering rows within df after manipulating it, e.g.: 
startSpiderSeqRDemo()
df1 <- searchAnywhere("trimkd")
df2 <- searchAnywhere("arabidopsis")
df <- rbind(df1, df2)
df <- orderDFAccessions(df)

}
\seealso{
Other Manipulation functions: \code{\link{filterByTermByAccessionLevel}},
  \code{\link{filterByTerm}}
}
\concept{Manipulation functions}
