% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{outputGenerator_acc}
\alias{outputGenerator_acc}
\title{Generate File Output (Accession)}
\usage{
outputGenerator_acc(df, ss = NULL, accession)
}
\arguments{
\item{df}{Data frame to output (unfilter)}

\item{ss}{A character vector with GSEs from verifySuperseries}

\item{accession}{A character vector with accessions 
for which the search was undertaken}
}
\value{
Nothing (generates relevant files)
}
\description{
Generate File Output (Accession)
}
\details{
Orders the df, replaces tabs (in characteristics_ch1), 
filters rows as appropriate, outputs files.

 @section Created files:
 \itemize{
     \item ChIP-Seq - if present
     \item RNA-Seq - if present
     \item OTHER - if present
 }
 All the files are made available as sample sheets (.csv and .Rda) 
 and database extracts (.csv, .Rda)
 
 @keywords internal
}
