% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{rbindUniqueCols}
\alias{rbindUniqueCols}
\title{Bind unique entries from two data frames disregarding some of the columns}
\usage{
rbindUniqueCols(x, y, disregard_columns)
}
\arguments{
\item{x, }{y Data frames to be merged (need to have the same columns; 
entries that are duplicated among the dfs will be removed from y)}

\item{disregard_columns}{Character vector with names of columns 
which will be ignored when finding unique rows 
of combined x and y data frame}
}
\value{
Data frame with all unique rows from x and y, 
when considered disregarding specified columns 
(only those rows of y that do not exist in x will be preserved)
}
\description{
Bind unique entries from two data frames disregarding some of the columns
}
\keyword{internal}
