% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{renameGSEColumns}
\alias{renameGSEColumns}
\title{Rename df columns derived from gse table to 'GSE_'}
\usage{
renameGSEColumns(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with modified column names
}
\description{
Rename df columns derived from gse table to 'GSE_'
}
\keyword{internal}
