% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{renameGSMColumns}
\alias{renameGSMColumns}
\title{Rename df columns derived from gsm table to 'GSM_'}
\usage{
renameGSMColumns(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with modified column names
}
\description{
Rename df columns derived from gsm table to 'GSM_'
}
\keyword{internal}
