% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{renameOTHColumns}
\alias{renameOTHColumns}
\title{Rename non-SRA/GEO columns to 'OTH_'}
\usage{
renameOTHColumns(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with modified column names
}
\description{
Rename non-SRA/GEO columns to 'OTH_'
}
\keyword{internal}
