% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{renameSRAColumns}
\alias{renameSRAColumns}
\title{Rename df columns derived from sra table to 'SRA_'}
\usage{
renameSRAColumns(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with modified column names
}
\description{
Rename df columns derived from sra table to 'SRA_'
}
