% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_Functions.R
\name{rerunSpideR}
\alias{rerunSpideR}
\title{Rerun SpideR query}
\usage{
rerunSpideR(file)
}
\arguments{
\item{file}{Query record generated at the time of query}
}
\value{
Rerun the query
}
\description{
This function allows for repeating the same query without the need to re-type the parameters over and over again. The main application of this function is to update results after the database files have been updated (i.e. to check whether there are any more results than last time the query was run).
}
\examples{
# rerunSpideR(file_name)

}
