% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_Anywhere.R
\name{searchAnywhereGSM}
\alias{searchAnywhereGSM}
\title{Search anywhere within gsm table of GEO database}
\usage{
searchAnywhereGSM(GSM_query, GEO_type)
}
\arguments{
\item{GSM_query}{String to search for}

\item{GEO_type}{Study type for filtering results (optional)}
}
\value{
Data frame with result data from whole GEO (gsm and gse tables)
}
\description{
Search anywhere within gsm table of GEO database
}
\keyword{internal}
