% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchForAccession.R
\name{searchForAccession}
\alias{searchForAccession}
\title{Search for Accessions}
\usage{
searchForAccession(acc_vector, file_output = FALSE, call_output = FALSE)
}
\arguments{
\item{acc_vector}{A character vector with accession numbers 
(of one type only)}

\item{file_output}{A logical indicating whether file output 
should be produced}

\item{call_output}{A logical indicating whether call output 
should be produced}
}
\value{
A data frame (and file outputs, as appropriate) 
     with matches to accession numbers
}
\description{
Search for Accessions
}
\examples{
startSpiderSeqRDemo()
df <- searchForAccession("SRP026280")

}
\seealso{
Other Workflow functions: 
\code{\link{addMissingSamples}()},
\code{\link{convertAccession}()},
\code{\link{filterByTermByAccessionLevel}()},
\code{\link{filterByTerm}()}

Other Core functions: 
\code{\link{convertAccession}()}
}
\concept{Core functions}
\concept{Workflow functions}
