% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession_Conversion_Functions.R
\name{searchForAccessionAcrossDBs}
\alias{searchForAccessionAcrossDBs}
\title{Search for accession across databases}
\usage{
searchForAccessionAcrossDBs(acc_vector, sra_columns, geo_columns)
}
\arguments{
\item{acc_vector}{A vector of accessions \strong{(all must belong to the same type)}}

\item{sra_columns}{A character vector with names of the columns to be returned from SRA}

\item{geo_columns}{A character vector with names of the columns to be returned from GEO}
}
\value{
A data frame with the results of the query
}
\description{
\code{searchForAccessionAcrossDBs} classifies accessions, searches for them in their original database (SRA or GEO), then, if conversion is possible, searches for corresponding accessions in the other database (SRA or GEO). If no conversion is possible, the columns from the corresponding database are returned empty (NAs).
}
\examples{
searchForAccessionAcrossDBs("GSE45530")

}
\keyword{internal}
