% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{searchSRA}
\alias{searchSRA}
\title{Test' to visualise the effects:
Test' to visualise the effects:}
\usage{
searchSRA(library_strategy, gene, antibody, cell_type, treatment, species,
  platform)
}
\description{
Test' to visualise the effects:
Test' to visualise the effects:
}
