% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession_Search_Functions.R
\name{searchSRAForAccession}
\alias{searchSRAForAccession}
\title{Search for SRA accessions}
\usage{
searchSRAForAccession(acc_vector, sra_columns)
}
\arguments{
\item{acc_vector}{A character vector with accession numbers}

\item{sra_columns}{A character vector with columns to be returned 
(from within sra table)}
}
\value{
A data frame with results
}
\description{
Search for SRA accessions
}
\examples{
# startSpiderSeqRDemo()
# searchSRAForAccession("SRP134708", "*")

}
\keyword{internal}
