% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{selectColumns_Overview}
\alias{selectColumns_Overview}
\title{Select columns with overview of sample information}
\usage{
selectColumns_Overview(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Original data frame containing only selected columns 
    (most relevant for getting an overview about the samples)
}
\description{
This is a shortcut function 
for \code{selectColumns(df, cols = listColumnSets()$Overview)}
}
\examples{
startSpiderSeqRDemo()
df <- searchForAccession("SRR2187611", file_output = FALSE)
df_sel <- selectColumns_Overview(df)

}
