% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_Functions.R
\name{selectColumnsAccession}
\alias{selectColumnsAccession}
\title{Select columns with accession information}
\usage{
selectColumnsAccession(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Original data frame containing only columns with accession information
}
\description{
This is a shortcut function for \code{selectColumns(df, cols = listColumnSets()$Accession)}
}
