% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_Option_Settings.R
\name{setSpiderSeqROption}
\alias{setSpiderSeqROption}
\title{Set options for functioning of the package}
\usage{
setSpiderSeqROption(name, value)
}
\arguments{
\item{name}{A character with the name of a variable}

\item{value}{Value to be assigned to the variable}
}
\value{
Nothing. Assign the value to a variable.
}
\description{
Set options for global behaviour of the package functions.
}
\section{Available options}{

\itemize{
    \item file_output - Logical - If TRUE, the search functions will produce 
    file outputs. If FALSE, no files will be created. Defaults to TRUE
   \item output_columns - Character vector - A set of columns to be used 
   for generating file outputs. 
   Defaults to one of the \code{listColumnSets()}.
   \item quiet - Whether any messages should be printed in the console
}
}

\examples{
setSpiderSeqROption("file_output", FALSE)
setSpiderSeqROption("output_columns", c("run_accession", "gsm"))
setSpiderSeqROption("output_columns", listColumnSets()$Accession)

}
