% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpideR.R
\docType{package}
\name{SpideR-package}
\alias{SpideR}
\alias{SpideR-package}
\title{SpideR: a tool for integration of Big Bio-Data}
\description{
SpideR integrates different sources of sequencing data and automates the search and retrieval of information.
}
\section{Things to include here ===*===}{


\itemize{
  \item How to access vignettes -  Example: `vignette("roxygen2", package = "roxygen2")`, `vignette("rd", package = "roxygen2")`, `vignette("namespace", package = "roxygen2")`
  \item Explicit links to information on setup
  \item List of databases accessed by SpideR
  \item Main functions (under development)
}
}

\section{SpideR functions}{

\itemize{
  \item \code{\link{searchForTerm}}
  \item \code{\link{searchSRA}}
  \item searchForAccession
  \item convertAccession
  \item reproduceResults (name to be changed)
}
}

\author{
\strong{Maintainer}: A. Sozanska \email{as2382@cam.ac.uk}

Authors:
\itemize{
  \item C. Fletcher
}

}
