% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Environmental_Settings.R
\docType{data}
\name{spiderEnv}
\alias{spiderEnv}
\title{SpideR environment}
\format{An object of class \code{environment} of length 0.}
\usage{
spiderEnv
}
\description{
SpideR environment
}
\keyword{datasets}
