% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_Option_Settings.R
\docType{data}
\name{SpiderSeqREnv}
\alias{SpiderSeqREnv}
\title{Environment with default values of the package options}
\format{
An object of class \code{environment} of length 6.
}
\usage{
SpiderSeqREnv
}
\description{
Environment for storing values determining certain aspects 
of package functioning.
To view the current value, please use \code{getSpiderSeqROption()}, 
to alter the current value, please use \code{setSpiderSeqROption()}.
}
\keyword{datasets}
