% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sra_demo}
\alias{sra_demo}
\title{DEMO extract from sra table (from sqlite file of SRAdb package)}
\format{
An object of class \code{data.frame} with 146 rows and 75 columns.
}
\usage{
sra_demo
}
\description{
Study accessions included in the data frame
\itemize{
 \item SRP134708
 \item DRP003157
 \item SRP061795
 \item SRP029758
 \item SRP076433
 \item SRP148363
 \item SRP062911
 
}
}
\details{
A data frame created as a DEMO for sra table from sqlite file 
of SRAdb package. 
Contains a few arbitrarily chosen complete study accessions.
}
\keyword{datasets}
