% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_Anywhere.R
\name{subsetSRAByAccessionLevel}
\alias{subsetSRAByAccessionLevel}
\title{Subset (by column) of a df based on SRA accession levels of interest}
\usage{
subsetSRAByAccessionLevel(df, acc_levels, add_run_accession = TRUE)
}
\arguments{
\item{df}{Data frame to be subset}

\item{acc_levels}{Accession levels of interest}
}
\value{
Data frame only with columns corresponding to accession levels of interest
}
\description{
Subset (by column) of a df based on SRA accession levels of interest
}
