% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSpiderSeqRConnections.R
\name{testSpiderSeqRConnections}
\alias{testSpiderSeqRConnections}
\title{Test SpiderSeqR Connections}
\usage{
testSpiderSeqRConnections()
}
\value{
Nothing. Prints information about the number of entries 
   in the key database tables.
}
\description{
Test SpiderSeqR Connections
}
\examples{
startSpiderSeqRDemo()
testSpiderSeqRConnections()

}
