% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unifying_Functions.R
\name{unifyDFFormat}
\alias{unifyDFFormat}
\title{Unify formats of data frames}
\usage{
unifyDFFormat(df)
}
\arguments{
\item{df}{Data frame to be processed}
}
\value{
Data frame in a unified format

Unifies the format of a data frame. The following transformations are made:
\itemize{
    \item Order columns
    \item Remove duplicate rows (\code{unique(df)})
    \item Order rows (and resetting row names)
    \item Convert data types (mostly to character, with some exceptions converted to integer)
}
}
\description{
Unify formats of data frames
}
\keyword{internal}
