% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{universalExtractor}
\alias{universalExtractor}
\title{Universal function for extracting from strings}
\usage{
universalExtractor(characteristics, key_words, sep_split, sep_collapse)
}
\arguments{
\item{characteristics}{Separable string (must be a string)}

\item{key_words}{A list of key words to be searched for (and removed); 
each of the k list levels contains synonyms for the same category}

\item{sep_split}{A character denoting splitting character}
}
\value{
A vector with:
\itemize{
    \item characteristics (original string)
    \item char_extract[1] - extract of key_word[[1]]
    \item ...
    \item char_extract[k] - extract of key_word[[k]]
}

Extracts parts of a separable string which contains key_words 
(without the key_words) and the remainder after all the subtractions
}
\description{
Universal function for extracting from strings
}
\keyword{internal}
