% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession_Functions.R
\name{vConditionVerifier}
\alias{vConditionVerifier}
\title{Verify matches to regular expressions}
\usage{
vConditionVerifier(regexpr_vector, x)
}
\arguments{
\item{regexpr_vector}{Vector with regular expressions (to be used by grepl)}

\item{x}{Character string to be checked for matches to regular expressions}
}
\value{
A logical vector with TRUE at positions where there was a match to at least one of the reqular expressions (OR operation)
}
\description{
Verify matches to regular expressions
}
\keyword{internal}
