% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{verifyColumns}
\alias{verifyColumns}
\title{Verify columns within a data frame}
\usage{
verifyColumns(df, column_list)
}
\arguments{
\item{column_list}{A character vector with column names to be checked}

\item{data}{frame (with named columns)}
}
\value{
Nothing but a printed message on the matching 
between colnames(df) and column_list

NOTE: Current version requires perfect matching (identical strings). 
Methods which acutally check the indices might be more flexible 
(e.g. case insensitive).
}
\description{
Verify columns within a data frame
}
\keyword{internal}
