% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{verifyConditions}
\alias{verifyConditions}
\title{Verify whether conditions are met in a data frame}
\usage{
verifyConditions(df, keywords, columns, prefixes, suffixes)
}
\arguments{
\item{df}{Data frame to be verified}

\item{keywords}{Vector of strings to be searched for}

\item{columns}{Vector with column names (within df) to be searched for}

\item{prefixes}{(Optional) Vector or list with prefixes to keywords 
(its first dimension needs to be the same as columns)}

\item{suffixes}{(Optional) Vector or list with suffixes to keywords
(its first dimension needs to be the same as columns)}
}
\value{
Logical vector (same length as df) 
with TRUE indicating matches to the conditions
}
\description{
Verify if the specified columns within the data frame 
match specified conditions. 
Match is performed using grepl (case insensitive, with perl set to TRUE).
Any given row will be returned as TRUE if ANY of its columns 
matched to ANY of the keywords with ANY of the prefixes 
and ANY of the suffixes.
}
\keyword{internal}
