% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession_Functions.R
\name{verifySimpleConditions}
\alias{verifySimpleConditions}
\title{Verify matches to regular expressions}
\usage{
verifySimpleConditions(regexpr_vector, x)
}
\arguments{
\item{regexpr_vector}{Vector with regular expressions (to be used by grepl)}

\item{x}{Character string to be checked for matches to regular expressions}
}
\value{
A logical vector with TRUE at positions where there was 
a match to at least one of the reqular expressions (OR operation)
}
\description{
A wrapper around grepl function (with default settings) 
performing an OR operation for matches of all the elements 
of the regexpr_vector elements (i.e. matching to one element is sufficient
for the value to be set to TRUE for that position)
}
\examples{
# verifySimpleConditions("^m$", c("m", "mm", "cm", "M"))

}
\keyword{internal}
