% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unifying_Functions.R
\name{checkValidColumns}
\alias{checkValidColumns}
\title{Check if column names of df are within allowed set}
\usage{
checkValidColumns(df)
}
\arguments{
\item{df}{Data frame to be checked}
}
\value{
Nothing - give an error or a warning

Current checks
\itemize{
    \item Argument is a data frame
    \item All columns within Valid Columns list
}
}
\description{
Check if column names of df are within allowed set
}
\keyword{internal}
