% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Console_Messages.R
\name{.vex}
\alias{.vex}
\title{Assign local variables to global environment (development mode)}
\usage{
.vex(name, value)
}
\arguments{
\item{name}{A string with the name of the local variable}

\item{value}{The local variable}
}
\value{
Nothing. Create the variable in the global environment
}
\description{
Assign local variables to global environment (development mode)
}
\keyword{internal}
