% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatabaseInformation.R
\name{getDatabaseInformation}
\alias{getDatabaseInformation}
\title{Get general information on database contents}
\usage{
getDatabaseInformation()
}
\value{
Data frame with requested information
}
\description{
\code{getDatabaseInformation} allows users to get insights into the contents of the database. Run the function and select the menu option with summary of interest.
}
\section{Available information}{

\itemize{
  \item \emph{Library_strategy counts in SRA database} - a dataframe with all the library strategies featuring in SRA and their counts
}
}

