% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{listValidColumns}
\alias{listValidColumns}
\title{List valid columns within the data frames}
\usage{
listValidColumns()
}
\value{
A list with column names grouped into categories
}
\description{
List valid columns within the data frames
}
\keyword{internal}
