% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{parQuery}
\alias{parQuery}
\title{Parametrised query for verifyMissingRuns()}
\usage{
parQuery(db_con, query, par_list)
}
\arguments{
\item{db_con, query, par_list}{Character vectors}
}
\value{
Data frame with results
}
\description{
Parametrised query for verifyMissingRuns()
}
\keyword{internal}
