% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_Functions.R
\name{searchForTerm}
\alias{searchForTerm}
\title{Search for samples matching criteria of interest}
\usage{
searchForTerm(library_strategy, gene = NULL, antibody = NULL,
  cell_type = NULL, treatment = NULL, species = NULL, platform = NULL,
  secondary_library_strategy = NULL)
}
\arguments{
\item{library_strategy}{Experimental method (e.g. RNA-Seq, ChIP-Seq). Only one library_strategy is allowed in a single query}

\item{gene}{A character vector with genes of interest (it is recommended to provide a few synonyms)}

\item{antibody}{A character vector with antibodies of interest (it is recommended to provide a few synonyms, some studies annotate their antibodies with trade names/symbols)}

\item{cell_type}{A character vector describing source types of interest (cell type, tissue, organ etc.)}

\item{treatment}{A character vector with keywords regarding treatment protocol}

\item{species}{A character vector with taxonomy IDs}

\item{platform}{A character vector with sequencing platforms}

\item{secondary_library_strategy}{Additional experimental method of interest filtered from the studies featured in search results}
}
\value{
Nothing. Creates a range of files with the query information and search results.
}
\description{
\code{searchForTerm} provides an automated framework for searching for samples matching a range of different criteria from the SRA database. It also supplements the sample information with data from GEO.
}
