% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_Functions.R
\name{searchForTerm}
\alias{searchForTerm}
\title{Search for samples matching criteria of interest}
\usage{
searchForTerm(library_strategy, gene = NULL, antibody = NULL,
  cell_type = NULL, treatment = NULL, species = NULL,
  platform = NULL, secondary_library_strategy = NULL)
}
\arguments{
\item{library_strategy}{Experimental method (e.g. RNA-Seq, ChIP-Seq). Only one library_strategy is allowed in a single query. To get a list of available library strategies, run \code{ getDatabaseInformation()}}

\item{gene}{A character vector with genes of interest (it is recommended to provide a few synonyms)}

\item{antibody}{A character vector with antibodies of interest (it is recommended to provide a few synonyms, some studies annotate their antibodies with trade names/symbols)}

\item{cell_type}{A character vector describing source types of interest (cell type, tissue, organ etc.)}

\item{treatment}{A character vector with keywords regarding treatment protocol}

\item{species}{A character vector with taxonomy IDs (e.g. "9606" for human)}

\item{platform}{A character vector with sequencing platforms}

\item{secondary_library_strategy}{Additional experimental method of interest filtered from the studies featured in search results}
}
\value{
Nothing. Creates a range of files with the query information and search results.
}
\description{
\code{searchForTerm} provides an automated framework for searching within SRA database for samples matching a range of different criteria (e.g. experimental method, tissue type). It also supplements the sample information with data from GEO (if available).
}
\section{Argument requirements}{

\strong{REQUIRED}: library_strategy AND at least one of: gene, antibody, cell_type or treatment

\strong{OPTIONAL}: species, platform, secondary_library_strategy
}

\section{Further information}{

For further information (especially on the output files) please refer to the package vignettes.
}

\examples{
searchForTerm(library_strategy = "RNA-Seq", gene = c("p53", "tp53"), species = "9606") 
st <- list(library_strategy="ChIP-Seq", gene="STAT1", antibody="STAT1", secondary_library_strategy = "RNA-Seq")
do.call(searchForTerm, st)


}
