% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{searchSRA}
\alias{searchSRA}
\title{An internal function for searchForTerm}
\usage{
searchSRA(SRA_library_strategy, gene, antibody, cell_type, treatment,
  species, platform)
}
\arguments{
\item{SRA_library_strategy, gene, antibody, cell_type, treatment, species, platform}{Character vectors with information to search within SRA}
}
\description{
An internal function for searchForTerm
}
\keyword{Test'}
\keyword{effects:}
\keyword{internal}
\keyword{the}
\keyword{to}
\keyword{visualise}
