% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_Option_Settings.R
\docType{data}
\name{spideREnv}
\alias{spideREnv}
\title{Default values of the package options}
\format{An object of class \code{environment} of length 3.}
\usage{
spideREnv
}
\description{
Default values of the package options
}
\keyword{datasets}
